Move lib/pam to session/pam
